+++ src/store/index.js
@@ -6,11 +6,11 @@ Vue.use(Vuex);
 export default new Vuex.Store({
   state: {
     drawerState: false,
-    basemapState: "Topo",
+    basemapState: "Streets",
     streamgageState: false,
     radarState: false,
     nfhlState: false,
-    allRPState: false,
+    allRPState: true,
     fwwState: false,
     currentZoomState: 4,
   },